Move add sprite and backdrop to their correct containers
